# backend/webAPI/server/common/api.yml
openapi: 3.0.1
info:
  title: webAPI
  description: Website API Server
  version: 1.0.0
servers:
  - url: /api/v1
security:
  - bearerAuth: []
paths:
...
  /users/{id}:
    parameters:
      - in: path
        name: id
        schema:
          type: integer
        required: true
        description: User ID
    get:
      tags:
        - Users
      summary: Get a user by ID
      responses:
        '200':
          description: A user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '404':
          description: User not found
    delete:
      tags:
        - Users
      summary: Delete a user
      responses:
        '200':
          description: The deleted user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '404':
          description: User not found
    put:
      tags:
        - Users
      summary: Update a user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
      responses:
        '200':
          description: The updated user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '404':
          description: User not found
...
components:
  schemas:
...
    User:
      type: object
      properties:
        ID:
          type: integer
          format: int64
          nullable: true
        Name:
          type: string
        Email:
          type: string
          format: email
        Phone:
          type: string
          nullable: true
        Address:
          type: string
          nullable: true
        CreatedAt:
          type: string
          format: date-time
          readOnly: true
        UpdatedAt:
          type: string
          format: date-time
          readOnly: true
        RoleID:
          type: integer
          format: int64
        FacultyID:
          type: integer
          format: int64
      required:
        - Name
        - Password
        - Email
        - UpdatedAt
        - RoleID
        - FacultyID
      title: User
...
  securitySchemes:
    bearerAuth: # arbitrary name for the security scheme
      type: http
      scheme: bearer
      bearerFormat: JWT